Move WalletConnect UI to home screen
